Use Slot B
I think in the future it would be convenient to build raw memory card
files for slot B, so Nintendont users can just copy it to their flash
drive instead of using Dolphin to move it to a .raw file. It would
probably be provided as part of a compressed archive.